% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_sample.R
\name{deconvolute_sample}
\alias{deconvolute_sample}
\title{deconvolution step}
\usage{
deconvolute_sample(
  sample.pat.path,
  reference,
  apply_weighting = T,
  verbose = F,
  num_of_inits = 1,
  max_iter = 100,
  n_threads = 1
)
}
\arguments{
\item{sample.pat.path}{}

\item{reference}{}

\item{num_of_inits}{numeric - how many random prior initializations to set for the EM approach. Default is 1 (uniform prior).}
}
\description{
deconvolution step
}
