% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{filter_pat}
\alias{filter_pat}
\title{filter_pat}
\usage{
filter_pat(
  pat,
  filter.noninf = T,
  filter.length = 3,
  filter.inf.length = 3,
  output.as.granges = F,
  verbose = F
)
}
\arguments{
\item{pat}{}

\item{filter.noninf}{}

\item{filter.length}{}

\item{filter.inf.length}{}

\item{output.as.granges}{}

\item{verbose}{}
}
\value{
filtered PAT ranges
}
\description{
filter_pat
}
