% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta.R
\name{fit_beta}
\alias{fit_beta}
\title{fit_beta
Fits a beta distribution to given methylation signals from sample pat file
Uses optim method to find mle and mme using empirical mean and variance}
\usage{
fit_beta(overlaps.list, pseudo = 1e-07, verbose = F)
}
\arguments{
\item{overlaps.list}{output from overlap_marker_pat}

\item{pseudo}{add pseudocount to avoid Inf/-Inf log values (default = 1e-7)}
}
\value{
beta params for each marker region using stats::optim
}
\description{
fit_beta
Fits a beta distribution to given methylation signals from sample pat file
Uses optim method to find mle and mme using empirical mean and variance
}
\examples{
\dontrun{
fit_beta(overlaps.list)
}
}
