% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta.R
\name{fit_beta_old}
\alias{fit_beta_old}
\title{fit_beta_old}
\usage{
fit_beta_old(overlaps.list, pf = 0.05)
}
\arguments{
\item{overlaps.list}{}
}
\value{
DEPRECATED - list of beta params using fitdistr
}
\description{
fit_beta_old
}
