% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pat_reads_at_marker_region.R
\name{get_pat_reads_at_marker_region}
\alias{get_pat_reads_at_marker_region}
\title{get_pat_reads_at_marker_region}
\usage{
get_pat_reads_at_marker_region(
  marker,
  marker.ind,
  pat.file = NULL,
  pat.dir = NULL,
  cell_type = NULL,
  filter.noninf = T,
  filter.length = 3,
  filter.inf.length = 3
)
}
\arguments{
\item{marker}{}

\item{pat.dir}{}

\item{cell_type}{}

\item{pat}{}
}
\value{
character vector of reads at specified marker region
}
\description{
get_pat_reads_at_marker_region
}
\examples{
\dontrun{
  marker <- use_reference()$marker
  pat <- System.file("")
  get_pat_reads_at_marker_region(marker, marker.ind = 5, pat.file = )
}


}
