% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_reference.R
\name{learn_reference}
\alias{learn_reference}
\title{learn_reference}
\usage{
learn_reference(marker.file, pat.dir, save.output = "", verbose = F)
}
\arguments{
\item{marker.file}{Path to marker file}

\item{pat.dir}{Path to directory containing reference pat files}

\item{save.output}{character - desired path to output reference, saved as serialized R object (.rds)}
}
\value{
reference object
}
\description{
Learn a reference set from given reference pat files of cell types of interest.
}
\examples{
\dontrun{
learn_reference(marker.file = "marker.txt", pat.dir = "data/ref/", save.output = "reference.rds")
}
}
