% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_fun.R
\name{likelihood_fun}
\alias{likelihood_fun}
\title{Calculate likelihood}
\usage{
likelihood_fun(psi, alpha, epsilon = 1e-99)
}
\arguments{
\item{psi}{matrix of psi values}

\item{alpha}{cell type proportions (same length as ncol(alpha))}

\item{epsilon}{pseudo-probability of psi in case psi = 0}
}
\value{
log-likelihood of input alpha value
}
\description{
Calculate likelihood
}
