% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyldeconvolveR-package.R
\docType{package}
\name{methyldeconvolveR-package}
\alias{methyldeconvolveR}
\alias{methyldeconvolveR-package}
\title{methyldeconvolveR: Reference-based deconvolution of tissues using DNA methylation}
\description{
Deconvolution of whole genome methylation sequencing data from bulk into individual cell types using a multinomial model. Inference of cell type proportions is performed using the expectation-maximization algorithm.
}
\author{
\strong{Maintainer}: Sidharth Singh Jain \email{sidharthsjain@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

Authors:
\itemize{
  \item Arthur Patrick McDeed \email{apm116@georgetown.edu} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
}

}
\keyword{internal}
