% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_marker_pat.R
\name{overlap_marker_pat}
\alias{overlap_marker_pat}
\title{Overlap marker and APT files}
\usage{
overlap_marker_pat(pat, marker, minoverlap = 1, type = "any", ...)
}
\arguments{
\item{pat}{Some description of PAT files}

\item{marker}{Some description of Marker files}

\item{minoverlap}{Minimum overlap of a read to a marker region}

\item{type}{Type of overlap, see \code[GenomicRanges]{}}
}
\description{
Merging the PAT and Marker files together
}
