% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_reference.R
\name{use_reference}
\alias{use_reference}
\title{use_reference}
\usage{
use_reference()
}
\value{
reference object
}
\description{
Function to access the published reference atlas described in McDeed et al.
}
\examples{

ref <- use_reference()
names(ref)


}
