% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_sample.R
\name{deconvolute_sample}
\alias{deconvolute_sample}
\title{deconvolution step}
\usage{
deconvolute_sample(
  sample.pat.path,
  reference,
  verbose = F,
  num_of_inits = 10,
  max_iter = 100,
  n_threads = 1
)
}
\arguments{
\item{sample.pat.path}{}

\item{reference}{}

\item{num_of_inits}{numeric - how many random prior initializations to set for the EM approach. Default is 1 (uniform prior).}
}
\value{
list of alpha from each initialization.
}
\description{
deconvolution step
}
\examples{
\dontrun{
ref = learn_reference(marker.file = "my_marker_file.txt", pat.dir = "directory/to/patfiles/")
deconvolute_sample_weighted(sample.pat.path = "sample_to_deconvolute.pat.gz", reference = ref)
}
}
