% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta.R
\name{fit_beta}
\alias{fit_beta}
\title{fit_beta}
\usage{
fit_beta(overlaps.list, epsilon = 1e-07)
}
\arguments{
\item{overlaps.list}{output from overlap_marker_pat}

\item{epsilon}{add pseudocount to avoid Inf/-Inf log values}
}
\value{
beta params for each marker region
}
\description{
fit_beta
}
