% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_reference.R
\name{learn_reference}
\alias{learn_reference}
\title{learn_reference}
\usage{
learn_reference(
  marker.file,
  pat.dir,
  save.output = "",
  verbose = T,
  n_threads = 1,
  split_reads = F,
  filter.inf.length = 3,
  filter.length = 3,
  filter.noninf = T,
  min.cell.param = NA,
  ...
)
}
\arguments{
\item{marker.file}{Path to marker file}

\item{pat.dir}{Path to directory containing reference pat files. See notes.}

\item{save.output}{character - desired path to output reference, saved as serialized R object (.rds)}
}
\value{
reference object
}
\description{
Learn a reference set from given reference pat files of cell types of interest.
}
\note{
Name formats of the reference PAT files should follow the following convention:
\itemize{
\item Files should end in .pat.gz (bgzipped PAT files outputted from wgbstools).
\item Files should have the name of the cell type (which perfectly matches the name in the marker file)
followed by an underscore, followed by any other names.
e.g. Bcell_1000_ExampleCellGeneratedInHouse.pat.gz
}
}
\examples{
\dontrun{
learn_reference(marker.file = "marker.txt", pat.dir = "data/ref/", save.output = "reference.rds")
}
}
