% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_marker_pat.R
\name{overlap_marker_pat}
\alias{overlap_marker_pat}
\title{Overlap marker and PAT files}
\usage{
overlap_marker_pat(pat, marker, split_reads = F, n_threads = 1)
}
\arguments{
\item{pat}{Some description of PAT files}

\item{marker}{Some description of Marker files}

\item{n_threads}{Number of threads to use (default = 1)}
}
\description{
Merging the PAT and Marker files together
}
