% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_marker}
\alias{read_marker}
\title{read_marker}
\usage{
read_marker(path = "data/Human_mixintest_top25.txt", linelimit = Inf)
}
\arguments{
\item{path}{}

\item{linelimit}{}
}
\value{
data.frame containing marker file contents
}
\description{
read_marker
}
\examples{
\dontrun{
read_marker(path = "inst_data/human_mixintest_top25.txt")
}
}
